+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Updated/Enhanced Rules and Bug Fixes
    - #318: SpaceAroundOperatorRule: Allow newline after elvis operator. (Russell Sanborn)
    - #310: Indentation rule: Incorrect level when outer class is None (i.e. script).
    - #313: Indentation rule: Nested classes declared within a condition report incorrect indentation.
+   - #332: SpaceBeforeOpeningBrace: Ignore opening brace preceded by opening ‘[’.
+   - #332: SpaceAfterClosingBrace: Ignore closing brace followed by closing ‘]’. Deprecate  and ignore *checkClosureMapEntryValue* property. 
 
 Framework and Infrastructure
    - #311: Add equivalent linux command line example. (Wilfred Hughes)